Added folders and organized files
