+++ inci/ingredient_group.py
@@ -1,4 +1,4 @@
-from cosing_db.db_utils import CosIng
+from inci_db.db_utils import CosIng
 
 
 class IngredientGroup(CosIng):
